# bondly-api/docs/swagger.yaml
basePath: /
definitions:
...
  dto.LoginRequest:
    properties:
      email:
        example: user@example.com
        type: string
      image_url:
        example: https://example.com/avatar.jpg
        type: string
      nickname:
        example: John Doe
        type: string
    required:
    - email
    - nickname
    type: object
...
  dto.SendCodeRequest:
    properties:
      email:
        example: user@example.com
        format: email
        type: string
    required:
    - email
    type: object
...
  dto.VerifyCodeRequest:
    properties:
      code:
        example: "123456"
        maxLength: 6
        minLength: 6
        type: string
      email:
        example: user@example.com
        format: email
        type: string
    required:
    - code
    - email
    type: object
...
  dto.WalletLoginRequest:
    properties:
      wallet_address:
        example: 0x1234567890123456789012345678901234567890
        type: string
    required:
    - wallet_address
    type: object
...
  response.Response-any:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
      success:
        type: boolean
    type: object
...
  response.ResponseComment:
    properties:
      code:
        example: 1000
        type: integer
      data: {}
      message:
        example: 操作成功
        type: string
      success:
        example: true
        type: boolean
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@bondly.com
    name: Bondly API Support
    url: https://bondly.com/support
  description: 这是Bondly项目的API文档，提供用户认证、区块链交互、内容管理和治理功能。
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://bondly.com/terms
  title: Bondly API
  version: "1.0"
paths:
  /api/v1/auth/code-status:
    get:
      consumes:
      - application/json
      description: 查询指定邮箱的验证码是否存在、剩余有效时间以及是否被限流。用于前端显示重发倒计时等功能。
      parameters:
      - description: 邮箱地址
        example: user@example.com
        format: email
        in: query
        name: email
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 邮箱参数缺失或格式错误
          schema:
            $ref: '#/definitions/response.Response-any'
      summary: 查询验证码状态
      tags:
      - 认证管理
  /api/v1/auth/login:
    post:
      consumes:
      - application/json
      description: 用户登录，如果用户不存在则自动创建新用户
      parameters:
      - description: 登录请求体
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 登录失败
          schema:
            $ref: '#/definitions/response.Response-any'
      summary: 用户登录
      tags:
      - 认证管理
  /api/v1/auth/send-code:
    post:
      consumes:
      - application/json
      description: 向指定邮箱发送6位数字验证码，用于用户身份验证。验证码有效期为10分钟，60秒内最多只能发送一次。
      parameters:
      - description: 发送验证码请求体
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.SendCodeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 请求参数错误或邮箱格式无效
          schema:
            $ref: '#/definitions/response.Response-any'
      summary: 发送邮箱验证码
      tags:
      - 认证管理
  /api/v1/auth/verify-code:
    post:
      consumes:
      - application/json
      description: 验证用户提交的6位数字验证码是否正确。验证成功后，验证码自动失效。
      parameters:
      - description: 验证码验证请求体
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.VerifyCodeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 请求参数错误或验证码不正确
          schema:
            $ref: '#/definitions/response.Response-any'
      summary: 验证邮箱验证码
      tags:
      - 认证管理
  /api/v1/auth/wallet-login:
    post:
      consumes:
      - application/json
      description: 用户仅通过钱包登陆，如果用户不存在则自动创建新用户
      parameters:
      - description: 钱包登录请求体
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.WalletLoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 钱包地址格式错误或登录失败
          schema:
            $ref: '#/definitions/response.Response-any'
      summary: 用户仅通过钱包登陆
      tags:
      - 认证管理
  /api/v1/blockchain/contract/{address}:
    get:
      consumes:
      - application/json
      description: 根据合约地址获取智能合约的详细信息，包括合约名称、符号、精度、总供应量等基本信息
      parameters:
      - description: 合约地址
        example: 0x1234567890abcdef1234567890abcdef12345678
        in: path
        maxLength: 42
        minLength: 42
        name: address
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 无效的合约地址格式
          schema:
            $ref: '#/definitions/response.Response-any'
      summary: 获取智能合约信息
      tags:
      - 区块链
  /api/v1/blockchain/stake:
    post:
      consumes:
      - application/json
      description: 质押指定数量的BONDLY代币到智能合约，获得治理权重和奖励。质押期间代币被锁定。
      parameters:
      - description: 质押请求体
        in: body
        name: request
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: 请求参数错误或余额不足
          schema:
            $ref: '#/definitions/response.Response-any'
      summary: 质押BONDLY代币
      tags:
      - 区块链
...
  /api/v1/comments:
    get:
      consumes:
      - application/json
      description: 分页获取评论列表
      parameters:
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 10
        description: 每页数量
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResponseComment'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ResponseAny'
      summary: 获取评论列表
      tags:
      - 评论管理
    post:
      consumes:
      - application/json
      description: 创建新的评论
      parameters:
      - description: 评论信息
        in: body
        name: comment
        required: true
        schema:
          $ref: '#/definitions/models.Comment'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.ResponseComment'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ResponseAny'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ResponseAny'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ResponseAny'
      security:
      - BearerAuth: []
      summary: 创建评论
      tags:
      - 评论管理
  /api/v1/comments/{id}:
    delete:
      consumes:
      - application/json
      description: 删除指定的评论
      parameters:
      - description: 评论ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResponseAny'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ResponseAny'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ResponseAny'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.ResponseAny'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ResponseAny'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ResponseAny'
      security:
      - BearerAuth: []
      summary: 删除评论
      tags:
      - 评论管理
    get:
      consumes:
      - application/json
      description: 根据ID获取评论详情
      parameters:
      - description: 评论ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResponseComment'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ResponseAny'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ResponseAny'
      summary: 获取评论详情
      tags:
      - 评论管理
    put:
      consumes:
      - application/json
      description: 更新指定评论的信息
      parameters:
      - description: 评论ID
        in: path
        name: id
        required: true
        type: integer
      - description: 更新的评论信息
        in: body
        name: comment
        required: true
        schema:
          $ref: '#/definitions/models.Comment'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResponseComment'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ResponseAny'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ResponseAny'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.ResponseAny'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ResponseAny'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ResponseAny'
      security:
      - BearerAuth: []
      summary: 更新评论
      tags:
      - 评论管理
  /api/v1/content:
    get:
      consumes:
      - application/json
      description: 获取平台上的内容列表，支持分页、分类筛选和关键词搜索。返回内容摘要、作者信息、点赞数等基本信息。
      parameters:
      - description: 页码，默认为1
        example: 1
        in: query
        minimum: 1
        name: page
        type: integer
      - description: 每页数量，默认为20
        example: 20
        in: query
        maximum: 100
        minimum: 1
        name: limit
        type: integer
      - description: 内容分类
        enum:
        - article
        - post
        - comment
        example: article
        in: query
        name: category
        type: string
      - description: 搜索关键词
        example: 区块链
        in: query
        name: keyword
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 内容列表
          schema:
            $ref: '#/definitions/response.Response-handlers_ContentListData'
      summary: 获取内容列表
      tags:
      - 内容管理
    post:
      consumes:
      - application/json
      description: 创建新的内容（文章、帖子等），需要提供标题、内容和类型。内容创建后默认为草稿状态。
      parameters:
      - description: 创建内容请求体
        in: body
        name: request
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/response.Response-any'
      summary: 创建新内容
      tags:
      - 内容管理
  /api/v1/content/{id}:
    delete:
      consumes:
      - application/json
      description: 删除指定的内容
      parameters:
      - description: 内容ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResponseAny'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ResponseAny'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ResponseAny'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.ResponseAny'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ResponseAny'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ResponseAny'
      security:
      - BearerAuth: []
      summary: 删除内容
      tags:
      - 内容管理
    get:
      consumes:
      - application/json
      description: 根据内容ID获取详细的内容信息，包括完整内容、作者信息、点赞/踩数、评论统计等。访问时会自动增加浏览次数。
      parameters:
      - description: 内容ID
        example: "1"
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 内容不存在或已被删除
          schema:
            $ref: '#/definitions/response.Response-any'
      summary: 获取内容详情
      tags:
      - 内容管理
    put:
      consumes:
      - application/json
      description: 更新指定内容的信息
      parameters:
      - description: 内容ID
        in: path
        name: id
        required: true
        type: integer
      - description: 更新的内容信息
        in: body
        name: content
        required: true
        schema:
          $ref: '#/definitions/models.Content'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResponseContent'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ResponseAny'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ResponseAny'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.ResponseAny'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ResponseAny'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ResponseAny'
      security:
      - BearerAuth: []
      summary: 更新内容
      tags:
      - 内容管理
...
  /api/v1/proposals:
    get:
      consumes:
      - application/json
      description: 分页获取提案列表
      parameters:
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 10
        description: 每页数量
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResponseProposal'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ResponseAny'
      summary: 获取提案列表
      tags:
      - 提案管理
    post:
      consumes:
      - application/json
      description: 创建新的提案
      parameters:
      - description: 提案信息
        in: body
        name: proposal
        required: true
        schema:
          $ref: '#/definitions/models.Proposal'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.ResponseProposal'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ResponseAny'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ResponseAny'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ResponseAny'
      security:
      - BearerAuth: []
      summary: 创建提案
      tags:
      - 提案管理
  /api/v1/proposals/{id}:
    delete:
      consumes:
      - application/json
      description: 删除指定的提案
      parameters:
      - description: 提案ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResponseAny'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ResponseAny'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ResponseAny'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.ResponseAny'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ResponseAny'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ResponseAny'
      security:
      - BearerAuth: []
      summary: 删除提案
      tags:
      - 提案管理
    get:
      consumes:
      - application/json
      description: 根据ID获取提案详情
      parameters:
      - description: 提案ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResponseProposal'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ResponseAny'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ResponseAny'
      summary: 获取提案详情
      tags:
      - 提案管理
    put:
      consumes:
      - application/json
      description: 更新指定提案的信息
      parameters:
      - description: 提案ID
        in: path
        name: id
        required: true
        type: integer
      - description: 更新的提案信息
        in: body
        name: proposal
        required: true
        schema:
          $ref: '#/definitions/models.Proposal'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResponseProposal'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ResponseAny'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ResponseAny'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.ResponseAny'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ResponseAny'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ResponseAny'
      security:
      - BearerAuth: []
      summary: 更新提案
      tags:
      - 提案管理
...
  /api/v1/transactions:
    get:
      consumes:
      - application/json
      description: 分页获取交易列表
      parameters:
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 10
        description: 每页数量
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResponseTransaction'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ResponseAny'
      summary: 获取交易列表
      tags:
      - 交易管理
    post:
      consumes:
      - application/json
      description: 创建新的交易记录
      parameters:
      - description: 交易信息
        in: body
        name: transaction
        required: true
        schema:
          $ref: '#/definitions/models.Transaction'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.ResponseTransaction'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ResponseAny'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ResponseAny'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ResponseAny'
      security:
      - BearerAuth: []
      summary: 创建交易
      tags:
      - 交易管理
  /api/v1/transactions/{id}:
    delete:
      consumes:
      - application/json
      description: 删除指定的交易
      parameters:
      - description: 交易ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResponseAny'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ResponseAny'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ResponseAny'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ResponseAny'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ResponseAny'
      security:
      - BearerAuth: []
      summary: 删除交易
      tags:
      - 交易管理
    get:
      consumes:
      - application/json
      description: 根据ID获取交易详情
      parameters:
      - description: 交易ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResponseTransaction'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ResponseAny'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ResponseAny'
      summary: 获取交易详情
      tags:
      - transaction
    put:
      consumes:
      - application/json
      description: 更新指定交易的信息
      parameters:
      - description: 交易ID
        in: path
        name: id
        required: true
        type: integer
      - description: 更新的交易信息
        in: body
        name: transaction
        required: true
        schema:
          $ref: '#/definitions/models.Transaction'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResponseTransaction'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ResponseAny'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ResponseAny'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ResponseAny'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ResponseAny'
      security:
      - BearerAuth: []
      summary: 更新交易
      tags:
      - 交易管理
  /api/v1/transactions/hash/{hash}:
    get:
      consumes:
      - application/json
      description: 根据交易哈希获取交易详情
      parameters:
      - description: 交易哈希
        in: path
        name: hash
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResponseTransaction'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ResponseAny'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ResponseAny'
      summary: 根据哈希获取交易
      tags:
      - 交易管理
  /api/v1/transactions/stats:
    get:
      consumes:
      - application/json
      description: 获取交易相关的统计数据
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResponseAny'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ResponseAny'
      summary: 获取交易统计
      tags:
      - 交易管理
  /api/v1/upload/image:
    post:
      consumes:
      - multipart/form-data
      description: 上传图片文件，支持jpg、jpeg、png、gif、webp格式，文件大小限制5MB
      parameters:
      - description: 图片文件
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: 文件格式错误或上传失败
          schema:
            $ref: '#/definitions/response.Response-any'
      summary: 上传图片文件
      tags:
      - 文件上传
  /api/v1/users:
    get:
      consumes:
      - application/json
      description: 分页获取用户列表
      parameters:
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 10
        description: 每页数量
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 获取用户列表成功
          schema:
            $ref: '#/definitions/response.Response-array_dto_UserResponse'
      summary: 获取用户列表
      tags:
      - 用户管理
    post:
      consumes:
      - application/json
      description: 创建新用户，支持Web2和Web3用户信息
      parameters:
      - description: 创建用户请求体
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 创建失败
          schema:
            $ref: '#/definitions/response.Response-any'
      summary: 创建新用户
      tags:
      - 用户管理
...
  /api/v1/users/{id}/custody-wallet:
    get:
      consumes:
      - application/json
      description: 获取指定用户的托管钱包地址和加密私钥
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 用户不存在
          schema:
            $ref: '#/definitions/response.Response-any'
      summary: 获取用户托管钱包信息
      tags:
      - 用户管理
...
schemes:
- http
- https
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"